+++ runtime/CHANGELOG.md
@@ -1,3 +1,11 @@
+### Version 9.10.0 - Giza
+- New storage and distribution runtime module
+- Renaming of working groups and adding new working group for distributor role
+- Enhancements to content directory module, supporting deletion of channels and videos.
+
+### Version 9.9.0 - Sumer - upgrade
+- Increase the max allowed working group mint capacity that can be set by council via proposals
+
 ### Version 9.7.0 - Sumer - runtime upgrade - May 27 2021
 - Introduced new content pallet the new content directory
 - Improved data_directory pallet